==> 01_oil reports/02_Preprocess_titles.ipynb <==
# load the data

import pandas as pd
df_1 = pd.read_pickle('sorted_all_reports.pkl')
df_2 = pd.read_pickle('contents_.pkl')

df_1.head()

df_2.head()

df = pd.merge(df_2, df_1,how='left', on=['date','content','title'])

df.head()

df.shape

df_2.shape

# preprocess title
## create title dataframe

def check_source(x):
    if '-' in x[1]:
        check_words = x[1].split('-')[0].split(' ')
    else:
        check_words = x[1].split(':')[0].split(' ')
    
    source_words=x[0].split(' ')
    
    flag=False
    for word in check_words:
        if word in source_words:
            flag=True

    if flag:
        if '-' in x[1]:
            new_title = ','.join(str(x) for x in x[1].split('-')[1:])
        if ':' in x[1]:
            new_title = ','.join(str(x) for x in x[1].split(':')[1:])
        else:
            new_title = x[1]
        return new_title

    else:
        return x[1]
    
    
    
    

df['new_title'] = df[['source', 'title']].apply(lambda x: check_source(x),axis=1)

## libraries

import pandas as pd
import numpy as np
 
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer, PorterStemmer
from string import punctuation
from collections import Counter
 
from collections import OrderedDict
import re
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)

## stop words, stemmers and lemmatizers

### add the source name in stop words

source = df.groupby('source').size().to_frame('counts').reset_index()
list_source = source['source'].tolist()

# source_head=set(",".join(str(x.split(' ')[0]) for x in list_source).\
#                 lower().\
#                 split(',')) # the word in the first of the name of source

source_head = set(",".join(str(x.split(' ')[1]) for x in list_source[-3:-1]).\
                lower().\
                split(','))

source_head = source_head.union(
                set(",".join(str(x.split(' ')[-1]).replace('(','').replace(')','')\
                for x in list_source).\
                lower().\
                split(','))
                )  # abbreviation of the source

### stop words and other words

porter = PorterStemmer()
wnl = WordNetLemmatizer() 
stop = set(stopwords.words('english'))
stop = stop.union({'delayed','jan','january','feb','february','mar','march','apr','april',
                   'may','june','jun','july','jul','aug','august','sep','september','oct','october','nov',
                  'november','dec','december',''})
stop = stop.union(source_head)
stop = stop.union({'daily','weekly','week','ltd','j.p.','report'})
# stop = set(stop)

# stop

## functions

def tokenizer(text):

    tokens_ = [word_tokenize(sent) for sent in sent_tokenize(text)]

    tokens = []
    for token_by_sent in tokens_:
        tokens += token_by_sent

#     tokens = [t for t in tokens if len(t) > 2] 
#     tokens = [t for t in tokens if not any(c.isdigit() for c in t)] 
    tokens = list(filter(lambda t: t.lower() not in stop, tokens))
    tokens = list(filter(lambda t: t not in punctuation, tokens))
    
     
    filtered_tokens = []
    for token in tokens:
        token = wnl.lemmatize(token) #convert to real word simple case
        if re.search('[a-zA-Z]', token): #regular expression
            filtered_tokens.append(token)

    filtered_tokens = list(map(lambda token: token.lower(), filtered_tokens))
    #map()是 Python 内置的高阶函数，它接收一个函数 f 和一个 list，并通过把函数 f 依次作用在 list 的每个元素上，得到一个新的 list 并返回。

    return filtered_tokens

def build_article_df(urls):
    articles = []
    for index, row in urls.iterrows(): #使用iterrows()对dataframe进行遍历
        try:
            data=row['new_title'].strip().replace("'", "") #如果strip()的参数为空，那么会默认删除字符串头和尾的空白字符(包括\n，\r，\t这些)。
            data = data.encode('ascii', 'ignore').decode('ascii')
            document = tokenizer(data)
          
            kw=" ".join(str(x) for x in document)
            articles.append((row['new_title'],kw,row['date'],row['new_source']))
            
        except Exception as e:
            print(e)
            pass

    article_df = pd.DataFrame(articles, columns=['title_to_handle','preprocess_title','date','new_source'])
    return article_df

## pre_title

# title_to_handle source = sorted_df.groupby('source').size().to_frame('counts').reset_index()
# title_list = df['new_title'].to_frame('title_to_handle')
title_list = df[['new_title','date','new_source']]

pre_title = build_article_df(title_list)
# pre_title

pre_title.shape

print(pre_title.iloc[0,0])
print(pre_title.iloc[0,1])

## Most frequent words

# list_ = pre_title['preprocess_title'].tolist()
# kw=" ".join(str(x) for x in list_)

# kw = kw.split(' ')

# Counter(kw).most_common(100)

## save pre_title

pre_title.to_pickle('preprocess_title.pkl')

# Extract features

## preprocess_title	--tfidf

# word count
pre_title['word_count'] = pre_title['preprocess_title'].apply(lambda x: len(str(x).split(" ")))
# pre_title.head(100)

pre_title['word_count'].quantile([0,0.10,0.2,0.3,0.5,0.8,0.9])

from sklearn.feature_extraction.text import TfidfVectorizer
tfidf = TfidfVectorizer(max_features=10, analyzer='word',ngram_range=(1,1))
tfidf_weights = tfidf.fit_transform(pre_title['preprocess_title']).toarray()

tfidf_vec = pd.DataFrame(tfidf_weights,columns=['title_tfidf_f{}'.format(i) for i in range(1,11)])
tfidf_vec.head()

tfidf_vec.to_pickle('title_tfidf_vec.pkl')

## Source vector

pre_title.groupby('new_source').size().to_frame().shape

from sklearn.feature_extraction.text import CountVectorizer
bow = CountVectorizer(max_features=114, analyzer='word',ngram_range=(1,1))
bow_weights = bow.fit_transform(pre_title['new_source']).toarray()

bow_weights.shape

source_vec = pd.DataFrame(bow_weights,columns=['source_bow_f{}'.format(i) for i in range(1,115)])
source_vec.head()

source_vec.to_pickle('source_bow_vec.pkl')
