==> 01_oil reports/03_Features.ipynb <==
# Extract date from original data

import pandas as pd
contents_ = pd.read_pickle('contents_.pkl')

# contents_.head()

original = contents_['date'].to_frame()

# POS tag

pos = pd.read_pickle('pos_vec_300.pkl')

pos = pd.merge(original,pos,left_index=True, right_index=True).set_index('date')

pos.shape

pos.to_pickle('./Features/feature_pos_vec_300.pkl')

# two_gram_vec

two_gram = pd.read_pickle("two_gram_vec_300.pkl")

two_gram = pd.merge(original,two_gram,left_index=True, right_index=True).set_index('date')

two_gram.tail()

two_gram.shape

two_gram.to_pickle('./Features/feature_two_gram_vec_300.pkl')

# tfidf_vec

tfidf = pd.read_pickle("tfidf_vec_300.pkl")

tfidf = pd.merge(original,tfidf,left_index=True, right_index=True).set_index('date')

# tfidf

tfidf.to_pickle('./Features/feature_tfidf_300.pkl')

# polarity and subjectivity	

p_s = contents_[['polarity','subjectivity','date']].set_index('date')

# p_s

p_s.to_pickle('./Features/feature_polarity_subjectivity.pkl')

# pretrained_Doc2Vec

Doc2Vec = pd.read_pickle("Doc2Vec_pretrained.pkl")

Doc2Vec = pd.merge(original,Doc2Vec,left_index=True, right_index=True).set_index('date')

# Doc2Vec

Doc2Vec.to_pickle('./Features/feature_Doc2Vec.pkl')

# title vector

titleVec = pd.read_pickle('title_tfidf_vec.pkl')

titleVec = pd.merge(original,titleVec,left_index=True, right_index=True).set_index('date')

titleVec.head()

titleVec.to_pickle('./Features/feature_titleVec.pkl')

# Source vector

sourceVec = pd.read_pickle('source_bow_vec.pkl')

sourceVec = pd.merge(original,sourceVec,left_index=True, right_index=True).set_index('date')

sourceVec.head()

sourceVec.to_pickle('./Features/feature_sourceVec.pkl')
