==> 01_oil reports/04_check time zone.ipynb <==
import pandas as pd
import os


def parse_text(file):

    out_dict = {}
    with open(folder+file, 'r',encoding='gb18030',errors='ignore') as f:
  
            text = f.read() 
            header = text.split('\n') # split text
            title = header[1]# get title
            
            if '[Delayed]' in title: # check if delayed
                if '-0-'in header[-10]:
                    zone = header[-10].split('-0- ')[1] # get date
                
#                     content = " ".join(str(x) for x in header[2:-10])
                else:
                    zone = header[0].split(' Date')[1].replace(': ','')
#                     content = " ".join(str(x) for x in header[2:-8])
            else:
                if '-0-'in header[-8]:
                    zone = header[-8].split('-0- ')[1]# get date
#                     content = " ".join(str(x) for x in header[2:-8])
                else:
                    zone = header[0].split(' Date')[1].replace(': ','')
#                     content = " ".join(str(x) for x in header[2:-7])
    
            out_dict['zone'] = zone

            
    return out_dict

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2013/'
files = os.listdir(folder)

files_dict_2013={}

for file in files:
    files_dict_2013[file]=parse_text(file)
# files_dict_2013

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2014/'
files = os.listdir(folder)

files_dict_2014={}

for file in files:
    files_dict_2014[file]=parse_text(file)
# files_dict_2014

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2015/'
files = os.listdir(folder)

files_dict_2015={}

for file in files:
    files_dict_2015[file]=parse_text(file)
# files_dict_2015

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2016/'
files = os.listdir(folder)

files_dict_2016={}

for file in files:
    files_dict_2016[file]=parse_text(file)
# files_dict_2016

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2017/'
files = os.listdir(folder)

files_dict_2017={}

for file in files:
    files_dict_2017[file]=parse_text(file)
# files_dict_2017

folder = 'C:/Users/Iris King/Desktop/MSc_FYP/original_reports/2018/'
files = os.listdir(folder)

files_dict_2018={}

for file in files:
    files_dict_2018[file]=parse_text(file)
# files_dict_2018

#2013
df_2013 = pd.DataFrame(columns=['zone'])
for name in files_dict_2013.keys():
    key = files_dict_2013[name]
    new = pd.DataFrame([key])
    df_2013 = pd.merge(df_2013, new, how='outer')

#2013
df_2013 = pd.DataFrame(columns=['zone'])
for name in files_dict_2013.keys():
    key = files_dict_2013[name]
    new = pd.DataFrame([key])
    df_2013 = pd.merge(df_2013, new, how='outer')
    
#2014
df_2014 = pd.DataFrame(columns=['zone'])
for name in files_dict_2014.keys():
    key = files_dict_2014[name]
    new = pd.DataFrame([key])
    df_2014 = pd.merge(df_2014, new, how='outer')
    
#2015
df_2015 = pd.DataFrame(columns=['zone'])
for name in files_dict_2015.keys():
    key = files_dict_2015[name]
    new = pd.DataFrame([key])
    df_2015 = pd.merge(df_2015, new, how='outer')
    
#2016
df_2016 = pd.DataFrame(columns=['zone'])
for name in files_dict_2016.keys():
    key = files_dict_2016[name]
    new = pd.DataFrame([key])
    df_2016 = pd.merge(df_2016, new, how='outer')

#2017
df_2017 = pd.DataFrame(columns=['zone'])
for name in files_dict_2017.keys():
    key = files_dict_2017[name]
    new = pd.DataFrame([key])
    df_2017 = pd.merge(df_2017, new, how='outer')
    
#2018
df_2018 = pd.DataFrame(columns=['zone'])
for name in files_dict_2018.keys():
    key = files_dict_2018[name]
    new = pd.DataFrame([key])
    df_2018 = pd.merge(df_2018, new, how='outer')

df_14 =  pd.merge(df_2013,df_2014,how='outer')
df_15 =  pd.merge(df_14,df_2015,how='outer')
df_16 =  pd.merge(df_15,df_2016,how='outer')
df_17 =  pd.merge(df_16,df_2017,how='outer')
df = pd.merge(df_17,df_2018,how='outer')

df.groupby('zone').size()

df_2013

df_2014
